SampleApp/openwrap/src/test/java/com/pubmatic/openwrap/models/POWApplicationInfoTest.java
/*
 * PubMatic Inc. ("PubMatic") CONFIDENTIAL
 * Unpublished Copyright (c) 2006-2020 PubMatic, All Rights Reserved.
 *
 * NOTICE:  All information contained herein is, and remains the property of PubMatic. The intellectual and technical concepts contained
 * herein are proprietary to PubMatic and may be covered by U.S. and Foreign Patents, patents in process, and are protected by trade secret or copyright law.
 * Dissemination of this information or reproduction of this material is strictly forbidden unless prior written permission is obtained
 * from PubMatic.  Access to the source code contained herein is hereby forbidden to anyone except current PubMatic employees, managers or contractors who have executed
 * Confidentiality and Non-disclosure agreements explicitly covering such access or to such other persons whom are directly authorized by PubMatic to access the source code and are subject to confidentiality and nondisclosure obligations with respect to the source code.
 *
 * The copyright notice above does not evidence any actual or intended publication or disclosure  of  this source code, which includes
 * information that is confidential and/or proprietary, and is a trade secret, of  PubMatic.   ANY REPRODUCTION, MODIFICATION, DISTRIBUTION, PUBLIC  PERFORMANCE,
 * OR PUBLIC DISPLAY OF OR THROUGH USE  OF THIS  SOURCE CODE  WITHOUT  THE EXPRESS WRITTEN CONSENT OF PUBMATIC IS STRICTLY PROHIBITED, AND IN VIOLATION OF APPLICABLE
 * LAWS AND INTERNATIONAL TREATIES.  THE RECEIPT OR POSSESSION OF  THIS SOURCE CODE AND/OR RELATED INFORMATION DOES NOT CONVEY OR IMPLY ANY RIGHTS
 * TO REPRODUCE, DISCLOSE OR DISTRIBUTE ITS CONTENTS, OR TO MANUFACTURE, USE, OR SELL ANYTHING THAT IT  MAY DESCRIBE, IN WHOLE OR IN PART.
 */
package com.pubmatic.openwrap.models;

import android.content.Context;
import android.content.pm.PackageInfo;
import android.content.pm.PackageManager;
import android.util.Log;

import junit.framework.Assert;

import org.junit.Test;
import org.junit.runner.RunWith;
import org.robolectric.RobolectricTestRunner;
import org.robolectric.annotation.Config;

import java.net.MalformedURLException;
import java.net.URL;

import androidx.test.core.app.ApplicationProvider;

@RunWith(RobolectricTestRunner.class)
@Config(manifest = Config.NONE, sdk = 19)
public class POWApplicationInfoTest {

    private static final String TAG = "PMAppInfoTest";

    @Test
    public void testAllClass() {
        Context context = ApplicationProvider.getApplicationContext();
        POWApplicationInfo appInfo = new POWApplicationInfo(context);

        // Test methods
        String appName = null, packageName = null;
        PackageManager manager = context.getPackageManager();
        PackageInfo info;
        try {
            info = manager.getPackageInfo(context.getPackageName(),
                    0);
            appName = info.applicationInfo.loadLabel(manager).toString();
            packageName = context.getPackageName();
        } catch (Exception e) {
            Log.e(TAG, e.getLocalizedMessage());
        }

        Assert.assertEquals(appName, appInfo.getName());
        Assert.assertEquals(packageName, appInfo.getPackageName());

        String inputText = "category1";
        appInfo.setCategories(inputText);
        Assert.assertEquals(inputText, appInfo.getCategories());

        inputText = "domaintest";
        appInfo.setDomain(inputText);
        Assert.assertEquals(inputText, appInfo.getDomain());

        inputText = "http://testurl.com/myapp";
        appInfo.setStoreURL(inputText);
        Assert.assertEquals(inputText, appInfo.getStoreURL().toString());
    }

}
